client: collapse Peer into RlpxPeer [no ci]
