+++ packages/client/src/net/peer/index.ts
@@ -2,5 +2,4 @@
  * @module net/peer
  */
 
-export * from './peer'
 export * from './rlpxpeer'
