+++ packages/client/src/sync/fetcher/bytecodefetcher.ts
@@ -12,7 +12,7 @@ import { keccak256 } from 'ethereum-cryptography/keccak'
 import { Fetcher } from './fetcher'
 import { getInitFecherDoneFlags } from './types'
 
-import type { Peer } from '../../net/peer'
+import type { RlpxPeer } from '../../net/peer'
 import type { FetcherOptions } from './fetcher'
 import type { Job, SnapFetcherDoneFlags } from './types'
 import type { BatchDBOp, DB } from '@ethereumjs/util'
@@ -255,7 +255,7 @@ export class ByteCodeFetcher extends Fetcher<JobTask, Uint8Array[], Uint8Array> 
   /**
    * Returns an idle peer that can process a next job.
    */
-  peer(): Peer | undefined {
+  peer(): RlpxPeer | undefined {
     return this.pool.idle((peer) => 'snap' in peer)
   }
